Amend package naming convention for bot layer.
