+++ api/src/main/java/me/longbow122/api/ApiApplication.java
@@ -6,12 +6,12 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ConfigurationPropertiesScan("me.longbow122.configuration.properties")
+@ConfigurationPropertiesScan("me.longbow122.bot.configuration.properties")
 @ComponentScan(basePackages = {
 	"me.longbow122.datamodel.repository",
-	"me.longbow122.service",
-	"me.longbow122.exception",
-	"me.longbow122.configuration",
+	"me.longbow122.bot.service",
+	"me.longbow122.bot.exception",
+	"me.longbow122.bot.configuration",
 	"me.longbow122.api.controller",
 })
 public class ApiApplication {
