+++ api/src/main/java/me/longbow122/api/controller/FormController.java
@@ -2,8 +2,8 @@ package me.longbow122.api.controller;
 
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
-import me.longbow122.dto.FormDTO;
-import me.longbow122.service.FormService;
+import me.longbow122.bot.dto.FormDTO;
+import me.longbow122.bot.service.FormService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.ResponseEntity;
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 @RestController
 @ComponentScan(basePackages = {
-	"me.longbow122.exception",
-	"me.longbow122.exception.exceptions"
+	"me.longbow122.bot.exception",
+	"me.longbow122.bot.exception.exceptions"
 })
 public class FormController {
