+++ bot/src/main/java/me/longbow122/bot/configuration/properties/DiscordConfigurationProperties.java
@@ -1,0 +1,7 @@
+package me.longbow122.bot.configuration.properties;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "discord")
+public record DiscordConfigurationProperties(String botToken) {
+}
