+++ bot/src/main/java/me/longbow122/bot/configuration/properties/FormConfigurationProperties.java
@@ -1,0 +1,8 @@
+package me.longbow122.bot.configuration.properties;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "form")
+public record FormConfigurationProperties(
+	long guildId) {
+}
