+++ bot/src/main/java/me/longbow122/bot/exception/exceptions/ChannelNotFoundException.java
@@ -1,0 +1,9 @@
+package me.longbow122.bot.exception.exceptions;
+
+import jakarta.persistence.EntityNotFoundException;
+
+public class ChannelNotFoundException extends EntityNotFoundException {
+	public ChannelNotFoundException(String message) {
+		super(message);
+	}
+}
