+++ bot/src/main/java/me/longbow122/bot/exception/exceptions/GuildNotFoundException.java
@@ -1,0 +1,9 @@
+package me.longbow122.bot.exception.exceptions;
+
+import jakarta.persistence.EntityNotFoundException;
+
+public class GuildNotFoundException extends EntityNotFoundException {
+	public GuildNotFoundException(String message) {
+		super(message);
+	}
+}
