+++ bot/src/main/java/me/longbow122/bot/exception/exceptions/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package me.longbow122.bot.exception.exceptions;
+
+import jakarta.persistence.EntityNotFoundException;
+
+public class UserNotFoundException extends EntityNotFoundException {
+	public UserNotFoundException(String message) {
+		super(message);
+	}
+}
