--- bot/src/main/java/me/longbow122/configuration/properties/DiscordConfigurationProperties.java
@@ -1,7 +1,0 @@
-package me.longbow122.configuration.properties;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "discord")
-public record DiscordConfigurationProperties(String botToken) {
-}
