--- bot/src/main/java/me/longbow122/configuration/properties/FormConfigurationProperties.java
@@ -1,8 +1,0 @@
-package me.longbow122.configuration.properties;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "form")
-public record FormConfigurationProperties(
-	long guildId) {
-}
