--- bot/src/main/java/me/longbow122/dto/validator/ListsMustHaveSameLength.java
@@ -1,22 +1,0 @@
-package me.longbow122.dto.validator;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ListsMustHaveSameLengthValidator.class)
-@Documented
-public @interface ListsMustHaveSameLength {
-	String message() default "Questions and Answers must be of the same length!";
-
-	Class<?>[] groups() default {};
-
-	Class<? extends Payload>[] payload() default {};
-
-	String firstList();
-
-	String secondList();
-}
