--- bot/src/main/java/me/longbow122/exception/exceptions/ChannelNotFoundException.java
@@ -1,9 +1,0 @@
-package me.longbow122.exception.exceptions;
-
-import jakarta.persistence.EntityNotFoundException;
-
-public class ChannelNotFoundException extends EntityNotFoundException {
-	public ChannelNotFoundException(String message) {
-		super(message);
-	}
-}
