--- bot/src/main/java/me/longbow122/exception/exceptions/GuildNotFoundException.java
@@ -1,9 +1,0 @@
-package me.longbow122.exception.exceptions;
-
-import jakarta.persistence.EntityNotFoundException;
-
-public class GuildNotFoundException extends EntityNotFoundException {
-	public GuildNotFoundException(String message) {
-		super(message);
-	}
-}
