--- bot/src/main/java/me/longbow122/exception/exceptions/UserNotFoundException.java
@@ -1,9 +1,0 @@
-package me.longbow122.exception.exceptions;
-
-import jakarta.persistence.EntityNotFoundException;
-
-public class UserNotFoundException extends EntityNotFoundException {
-	public UserNotFoundException(String message) {
-		super(message);
-	}
-}
