+++ bot/src/test/java/me/longbow122/bot/configuration/TestConfiguration.java
@@ -1,0 +1,8 @@
+package me.longbow122.bot.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+@Configuration(proxyBeanMethods = false)
+@Profile("test")
+public class TestConfiguration {}
