--- bot/src/test/java/me/longbow122/configuration/TestConfiguration.java
@@ -1,8 +1,0 @@
-package me.longbow122.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
-@Configuration(proxyBeanMethods = false)
-@Profile("test")
-public class TestConfiguration {}
